Committing and pushing this project to GitHub, to get it off of my desktop.
I was originally working on this project with Sideways Eight Productions,
but it fell by the wayside, and now I'm not interested in it at all
anymore. OpenGL will always be an option for me, but learning it in a
project this big isn't ideal.

I don't even have a clear idea of what I wanted anymore, so I'm officially
deprecating this project.